requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
django-js-asset==2.0.0
django-markdownx==4.0.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
importlib-metadata==6.6.0
Markdown==3.4.3
Pillow==9.5.0
pytz==2022.7.1
sqlparse==0.4.3
